+++ libs/execution/src/lib/execution.spec.ts
@@ -1,0 +1,7 @@
+import { execution } from './execution.js';
+
+describe('execution', () => {
+  it('should work', () => {
+    expect(execution()).toEqual('execution');
+  });
+});