+++ libs/market/src/lib/market.spec.ts
@@ -1,0 +1,7 @@
+import { market } from './market.js';
+
+describe('market', () => {
+  it('should work', () => {
+    expect(market()).toEqual('market');
+  });
+});