+++ libs/onchain/src/lib/onchain.spec.ts
@@ -1,0 +1,7 @@
+import { onchain } from './onchain.js';
+
+describe('onchain', () => {
+  it('should work', () => {
+    expect(onchain()).toEqual('onchain');
+  });
+});