+++ libs/replay/src/lib/replay.spec.ts
@@ -1,0 +1,7 @@
+import { replay } from './replay.js';
+
+describe('replay', () => {
+  it('should work', () => {
+    expect(replay()).toEqual('replay');
+  });
+});