+++ libs/risk/src/lib/risk.spec.ts
@@ -1,0 +1,7 @@
+import { risk } from './risk.js';
+
+describe('risk', () => {
+  it('should work', () => {
+    expect(risk()).toEqual('risk');
+  });
+});