+++ libs/signals/src/lib/signals.spec.ts
@@ -1,0 +1,7 @@
+import { signals } from './signals.js';
+
+describe('signals', () => {
+  it('should work', () => {
+    expect(signals()).toEqual('signals');
+  });
+});