+++ libs/solana-hard-state/src/lib/solana-hard-state.spec.ts
@@ -1,0 +1,7 @@
+import { solanaHardState } from './solana-hard-state.js';
+
+describe('solanaHardState', () => {
+  it('should work', () => {
+    expect(solanaHardState()).toEqual('solana-hard-state');
+  });
+});