+++ libs/strategies/src/lib/strategies.spec.ts
@@ -1,0 +1,7 @@
+import { strategies } from './strategies.js';
+
+describe('strategies', () => {
+  it('should work', () => {
+    expect(strategies()).toEqual('strategies');
+  });
+});